requirements.txt
PyQt6
PyQt6-WebEngine
Flask
Pillow
